requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose
python-dotenv
pydantic>=2,<3
email-validator

passlib[bcrypt]==1.7.4
bcrypt==3.2.2
